+++ src/main/python/plotlyst/service/persistence.py
@@ -299,15 +299,15 @@ def delete_character(novel: Novel, character: Character, forced: bool = False) -
         repo.delete_character(novel, character)
 
         char_id = character.id
-        removed_conflicts = []
-        for conflict in novel.conflicts:
-            if conflict.character_id == char_id or conflict.conflicting_character_id == char_id:
-                removed_conflicts.append(conflict)
-        for conf in removed_conflicts:
-            novel.conflicts.remove(conf)
-        if removed_conflicts:
-            repo.update_novel(novel)
-        removed_conflicts = [x.id for x in removed_conflicts]
+        # removed_conflicts = []
+        # for conflict in novel.conflicts:
+        #     if conflict.character_id == char_id:
+        #         removed_conflicts.append(conflict)
+        # for conf in removed_conflicts:
+        #     novel.conflicts.remove(conf)
+        # if removed_conflicts:
+        #     repo.update_novel(novel)
+        # removed_conflicts = [x.id for x in removed_conflicts]
 
         for scene in novel.scenes:
             update_scene = False
@@ -321,17 +321,10 @@ def delete_character(novel: Novel, character: Character, forced: bool = False) -
                     break
             for agenda in scene.agency:
                 if agenda.character_id == char_id:
-                    agenda.conflict_references.clear()
-                    agenda.goal_references.clear()
+                    agenda.conflicts.clear()
                     agenda.motivations.clear()
                     update_scene = True
                     continue
-                if removed_conflicts:
-                    before = len(agenda.conflict_references)
-                    agenda.conflict_references[:] = [x for x in agenda.conflict_references
-                                                     if x.conflict_id not in removed_conflicts]
-                    if before != len(agenda.conflict_references):
-                        update_scene = True
 
             if update_scene:
                 repo.update_scene(scene)
